Print the word in the correct order.
The input will be an array of strings like ["J>A", "A>G", "G>U"]. Each string will always have 3 characters in length, the middle character will always be '>'. There will be no repetition of characters in the resultant string. For J>A, it means that J should come ahead of A in the resulting string. Likewise, A>G means A should come before G in the resultant string. The resultant string For ["J>A", "A>G", "G>U"] is: JAGU.
Example-1
-------------------------------
Input:["J>A", "A>G", "G>U"]
Output: JAGU

Example-2
-------------------------------
Input:["A>G","J>A", "G>U"]
Output: JAGU